add component for UserMenu
